[support] Add custom checks support
