+++ ek.go
@@ -21,7 +21,7 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // VERSION is current ek package version
-const VERSION = "12.108.2"
+const VERSION = "12.109.0"
 
 // ////////////////////////////////////////////////////////////////////////////////// //
