+++ support/support_test.go
@@ -54,6 +54,7 @@ func (s *SupportSuite) TestCollect(c *C) {
 	c.Assert(i.WithRevision("1234567"), NotNil)
 	c.Assert(i.WithPackages([]Pkg{Pkg{"test", "test-1.2.3"}}), NotNil)
 	c.Assert(i.WithApps(App{"test", "1.2.3"}), NotNil)
+	c.Assert(i.WithChecks(Check{CHECK_OK, "Test", "Test message"}), NotNil)
 	c.Assert(i.WithEnvVars("", "TEST", "TERM", "CI"), NotNil)
 	c.Assert(i.WithNetwork(&NetworkInfo{PublicIP: "192.168.1.1"}), NotNil)
 	c.Assert(i.WithFS([]FSInfo{FSInfo{}, FSInfo{"/", "/dev/vda1", "ext4", 1000, 10000}}), NotNil)
@@ -66,6 +67,7 @@ func (s *SupportSuite) TestNil(c *C) {
 	c.Assert(i.WithRevision(""), IsNil)
 	c.Assert(i.WithPackages(nil), IsNil)
 	c.Assert(i.WithApps(App{}), IsNil)
+	c.Assert(i.WithChecks(Check{}), IsNil)
 	c.Assert(i.WithEnvVars(""), IsNil)
 	c.Assert(i.WithNetwork(nil), IsNil)
 	c.Assert(i.WithFS(nil), IsNil)
